RTCDataChannel close event does not fire if RTCPeerConnection closes rdar://165617848 https://bugs.webkit.org/show_bug.cgi?id=303052
Reviewed by NOBODY (OOPS!).

Since closing the peer connection backend is severing the state updates of the data channel backend, we add the logic to close the data channel from RTCPeerConnection::close.
For that reason, we keep a list of channel identifiers in RTCPeerConnection.

We then call RTCDataChannelHandlerClient::peerConnectionIsClosing for each of this identifier.
This will do the following:
- If the data channel lives in main thread, it will call RTCDataChannel::peerConnectionIsClosing() directly.
- If the data channel lives in another process, it will call RTCDataChannel::peerConnectionIsClosing() directly.
- If the data channel lives in a worker, it will call RTCDataChannel::peerConnectionIsClosing() after hopping to the service worker thread.
- If the data channel is being transferred, it will call RTCDataChannel::peerConnectionIsClosing() which will make the newly transferred data channel moving to closed state.

Covered by added test.